Update some more system tests, add comments to the code etc.
